Add record update method to AdaloTableService
The commit adds a new method 'PutAsync' to the AdaloTableService. This method allows updates to a record in the Adalo database using an update action. Furthermore, it adjusts the WeeksService and DetailledDayDto to accommodate changes in data structuring and also adds a new route to SoleyaProjectRegistration for updating reservation details.