Refactor Adalo and API Key services for improved encapsulation
Adalo service declaration and return type are changed to provide better encapsulation. Service exception handling is also modified to ensure required services are present. Internal access modifiers are used in the API key authentication handler and extensions for encapsulation and better object orientation, and to prevent erroneous access of the classes. Additionally, a new file AdaloServiceCollection.cs is created for further services manipulation.