Add cache duration option to Adalo table
This update enables setting a custom cache duration when adding a table to Adalo through the AddTable method in AdaloAppOptions.cs. The AdaloTableCacheService.cs has been updated to use the new cache duration setting when refreshing and updating the table cache. The AdaloTableOptions.cs file has been updated to include the CacheDuration property.