Add cache controller generation option and improve caching mechanism
The commit introduces an option to generate cache controllers when adding a table in AdaloAppOptions. Moreover, improve the caching mechanism in AdaloTableCacheService to separate table and record caching. Add a refresh mechanism for both the record and table cache. This update provides more control over cache management and enhances the performance of the system.